backend/requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-community==0.3.7

# Database
sqlalchemy==2.0.35
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# HTTP Client
httpx==0.28.0
aiohttp==3.11.7

# Utilities
python-multipart==0.0.17
